Update README.md
- In  “Step 2: Writing the Smart Contract”:
a) I’ve added an event called ValueSet that is emitted whenever the value variable is set. 
b) I've also added error handling to the setValue function using the require keyword to check that the value being set is not empty. If the value is empty, the function will revert and the error message "Value cannot be empty" will be returned.


-Added relevant comments to the code as they provide a brief overview of what the contract does and explain the purpose of each function and variable.

- Applied code indentation.

- In “Step 4: Building the Frontend Interface”
a) Used constants for contractAddress and contractABI instead of hardcoded values. This will make it easier to modify them in the future if needed.
b) Added error handling for when a user tries to setValue without entering a new value.
c) Added loading indicators for when a transaction is being processed and when getValue is being called.


- Added table of contents.
-Added relevant links to the source.
- Did paraphrasing.
- Added punctuations wherever, required.
- Corrected grammatical errors.
- Added the conclusion so that the tutorial is complete.
- Made code snippets.
- Formatted the entire document by adding/removing ample spaces between paragraphs so that it looks well-structured.